perf: cache classification vocabulary ordering
